pyproject.toml
[project]
name = "userlab"
version = "0.1.0"
description = "User management teaching platform"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "streamlit>=1.47.1",
]

